Answer the review findings on naming, contracts, and suite shape
