+++ test/shared/turso-api/provider.test.ts
@@ -1,0 +1,253 @@
+import { expect } from "@std/expect";
+import { it as test } from "@std/testing/bdd";
+import { tursoDbProvider as tursoApi } from "#shared/turso-api.ts";
+import { testCreateDatabaseReturnsErrorOn403 } from "#test-utils/builder-mocks.ts";
+import { describeWithEnv } from "#test-utils/db.ts";
+import { stubFetch } from "#test-utils/fetch-stub.ts";
+import { withMocks } from "#test-utils/mocks.ts";
+import { createReplyingWith, TEST_TURSO_ENV } from "./fixtures.ts";
+
+/**
+ * Stub fetch with standard Turso URL routing: /databases → db JSON, /auth/tokens → JWT.
+ * Pass `onRequest` to capture URL/init for inspection. Returns on unmatched URLs with 500.
+ */
+const stubTursoFetch = (
+  dbName: string,
+  dbId: string,
+  jwt: string,
+  onRequest?: (url: string, init?: RequestInit) => void,
+) =>
+  stubFetch((url, init) => {
+    onRequest?.(url, init);
+    if (url.includes("/databases") && !url.includes("/auth")) {
+      return new Response(
+        JSON.stringify({
+          database: {
+            DbId: dbId,
+            Hostname: `${dbName}.turso.io`,
+            Name: dbName,
+          },
+        }),
+      );
+    }
+    if (url.includes("/auth/tokens")) {
+      return new Response(JSON.stringify({ jwt }));
+    }
+    return new Response("unexpected", { status: 500 });
+  });
+
+/** Returns a `stubTursoFetch` `onRequest` callback that captures the create-database POST body into `out.body`. */
+const captureCreateBody =
+  (out: { body?: unknown }) => (url: string, init?: RequestInit) => {
+    if (url.includes("/databases") && !url.includes("/auth")) {
+      out.body = JSON.parse(init?.body as string);
+    }
+  };
+
+describeWithEnv("turso-api", { env: TEST_TURSO_ENV }, () => {
+  test("createDatabase calls create and token endpoints", async () => {
+    const fetchCalls: string[] = [];
+
+    await withMocks(
+      () =>
+        stubTursoFetch("my-site", "db_test", "jwt_token", (url) =>
+          fetchCalls.push(url),
+        ),
+      async () => {
+        const result = await tursoApi.createDatabase("My Site");
+
+        expect(result.ok).toBe(true);
+        if (result.ok) {
+          expect(result.value.dbUrl).toBe("libsql://my-site.turso.io");
+          expect(result.value.dbToken).toBe("jwt_token");
+          expect(result.value.dbId).toBe("db_test");
+        }
+
+        expect(fetchCalls.length).toBe(2);
+      },
+    );
+  });
+
+  test("createDatabase POSTs name and group to the create endpoint", async () => {
+    const out: { body?: unknown } = {};
+    await withMocks(
+      () => stubTursoFetch("x", "db_x", "j", captureCreateBody(out)),
+      async () => {
+        await tursoApi.createDatabase("x");
+        expect(out.body).toEqual({ group: "default", name: "x" });
+      },
+    );
+  });
+
+  test("createDatabase slugifies the name before sending to Turso API", async () => {
+    const slug: { body?: unknown } = {};
+    await withMocks(
+      () => stubTursoFetch("my-site", "db_slug", "j", captureCreateBody(slug)),
+      async () => {
+        await tursoApi.createDatabase("My Site");
+        expect(slug.body).toEqual({ group: "default", name: "my-site" });
+      },
+    );
+  });
+
+  test("createDatabase uses Bearer token auth header", async () => {
+    const authHeaders: string[] = [];
+
+    await withMocks(
+      () =>
+        stubTursoFetch("h", "db_h", "t", (_url, init) => {
+          const auth = (init?.headers as Record<string, string>)?.Authorization;
+          if (auth) authHeaders.push(auth);
+        }),
+      async () => {
+        await tursoApi.createDatabase("h");
+        expect(authHeaders.every((h) => h === "Bearer test-turso-token")).toBe(
+          true,
+        );
+        expect(authHeaders.length).toBeGreaterThan(0);
+      },
+    );
+  });
+
+  test("createDatabase uses org and db name in URLs", async () => {
+    const fetchedUrls: string[] = [];
+
+    await withMocks(
+      () =>
+        stubTursoFetch("url-db", "db_url", "tok", (url) =>
+          fetchedUrls.push(url),
+        ),
+      async () => {
+        await tursoApi.createDatabase("url-db");
+        expect(fetchedUrls.some((u) => u.includes("/myorg/"))).toBe(true);
+        expect(fetchedUrls.some((u) => u.includes("/url-db/"))).toBe(true);
+        expect(
+          fetchedUrls.some((u) => u.includes("authorization=full-access")),
+        ).toBe(true);
+      },
+    );
+  });
+
+  test("createDatabase returns error when create endpoint fails", async () => {
+    await testCreateDatabaseReturnsErrorOn403(tursoApi);
+  });
+
+  test("createDatabase returns error when create endpoint fails with JSON error", async () => {
+    await withMocks(
+      () =>
+        stubFetch(
+          new Response(JSON.stringify({ error: "quota exceeded" }), {
+            status: 422,
+          }),
+        ),
+      async () => {
+        const result = await tursoApi.createDatabase("Bad");
+        expect(result.ok).toBe(false);
+        if (!result.ok) {
+          expect(result.error).toContain("Create database failed (422)");
+          expect(result.error).toContain("quota exceeded");
+        }
+      },
+    );
+  });
+
+  test("createDatabase returns error when token endpoint fails", async () => {
+    await withMocks(
+      () =>
+        stubFetch((url) => {
+          if (url.includes("/databases") && !url.includes("/auth")) {
+            return new Response(
+              JSON.stringify({
+                database: {
+                  DbId: "db_tok_fail",
+                  Hostname: "t.turso.io",
+                  Name: "t",
+                },
+              }),
+            );
+          }
+
+          return new Response(JSON.stringify({ message: "Unauthorized" }), {
+            status: 401,
+          });
+        }),
+      async () => {
+        const result = await tursoApi.createDatabase("t");
+        expect(result.ok).toBe(false);
+        if (!result.ok) {
+          expect(result.error).toContain(
+            "Generate database token failed (401)",
+          );
+          expect(result.error).toContain("Unauthorized");
+        }
+      },
+    );
+  });
+
+  test("createDatabase constructs libsql:// URL from hostname", async () => {
+    await withMocks(
+      () => stubTursoFetch("my-app", "db_turso123", "jwt_abc123"),
+      async () => {
+        const result = await tursoApi.createDatabase("My App");
+        expect(result.ok).toBe(true);
+        if (result.ok) {
+          expect(result.value.dbUrl).toMatch(/^libsql:\/\//);
+          expect(result.value.dbUrl).toContain("my-app.turso.io");
+        }
+      },
+    );
+  });
+
+  test("createDatabase refuses a database host that is not a bare hostname", async () => {
+    // A host only round-trips through the URL parser when it is bare:
+    // lowercase, no port, no spaces. Anything else means the platform is not
+    // pointing at a plain database host, so the schema must reject it.
+    const jwt = JSON.stringify({ jwt: "j" });
+    const refusedHosts: [name: string, hostname: string][] = [
+      ["db_port", "db.example.com:8080"],
+      ["db_case", "DB.Example.COM"],
+      ["db_space", "Not A Hostname"],
+    ];
+    for (const [name, hostname] of refusedHosts) {
+      await withMocks(
+        () =>
+          stubFetch(
+            createReplyingWith(
+              { DbId: name, Hostname: hostname, Name: "ported" },
+              jwt,
+            ),
+          ),
+        async () => {
+          const result = await tursoApi.createDatabase("Ported");
+          expect(result.ok, hostname).toBe(false);
+          if (!result.ok) {
+            expect(result.error).toContain(
+              "Create database returned an invalid response",
+            );
+          }
+        },
+      );
+    }
+  });
+
+  test("createDatabase names the token operation when its reply is not JSON", async () => {
+    await withMocks(
+      () =>
+        stubFetch(
+          createReplyingWith(
+            { DbId: "d", Hostname: "t.turso.io", Name: "halfbroken" },
+            "not json",
+          ),
+        ),
+      async () => {
+        const result = await tursoApi.createDatabase("HalfBroken");
+        expect(result.ok).toBe(false);
+        if (!result.ok) {
+          expect(result.error).toContain(
+            "Generate database token returned invalid JSON",
+          );
+        }
+      },
+    );
+  });
+});
